requirements.txt
azure-ai-textanalytics
azure-ai-translation-text==1.0.0b1